# Optimization : globpath using godirwalk