Hello,

### What it does

This PR adds 2 simple optimizations to globpath internal library :
- returns file names rather than fileInfo (-> prevents us stat-ing found files, as fileInfo is most of the times discarded by the caller)
- replaces filepath.walk with godirwalk library. The latter is more efficient (less `stat` calls).

### Not sorting results anymore

I used `unsorted : true` option of godirwalk. It changes the order of recursion (the option name is quite explicit), and thus the order of the results provided by telegraf plugins. As there is no contract about the order of results with telegraf users, i don't consider this move as a breaking change. But this (small) optimization could be reverted if you disagree with that point.

### Results

While it is an optimization on the paper, results are not striking. Simple `time` results with and without the optimization, run on path relative to telegraf source tree :

| **plugin** | **path** | **master** | **optimization**
|--------|----|---|------------- |
|file|/**.csv|0,346s|0,203s|
|filestat (no md5) | /**.go|0,539s|0,463s|
|filestat (no md5) | /** |0,775s|0,816s|

I don't know why in the last case the 'optimized' code performs less than the non-optimized one. Maybe because filestat `stat`s every results, so the first optimization doesn't apply.

### Required for all PRs:

- [x] Signed [CLA](https://influxdata.com/community/cla/).
- [pointless] Associated README.md updated.
- [x] Has appropriate unit tests.